+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* Removed Flutter as a dependency
+* Added characteristic properties
+* Added toLowerCase to every call that requires a UUID, because the web api expects only lower case UUIDS.
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Basic support for web bluetooth api